Add jitter rejection to images
This is added based on the has_errors and is_silent flag that
accompanies all images. As these flags do not work for the Sphinx
simulator, this is only done when the images are from a real-life
camera. This fault with the simulator does not really matter since the
images coming from it typically contain very little jitter anyways.