Fix dnncv pos. to be in NED, add error handling to drone interface
The DNNCV position is now properly in NED coordinates. This is not
the case for the ground truth as this still has ENU coordinates for
the x- and y-axis (will fix this in a future commit).

Also added some error handling to the drone interface as it
sometimes throws errors. The handling basically just replaced the
value the it failed to get from the drone with an appropriate one,
so if it fails to get a gps fix status it assumes no fix, and if it
failes to get the battery health it assumes it's good.