Fix bug with changed x and y-axis in ground truth
The bug was that the x and y-axis were changed so that forward was
the y-axis and right was the x-axis. This was caused by an incorrect
rotation in the in the rotation matrix, which was caused because the
angle coming from the sphinx-ros-bridge is defined positive counter
clockwise. All angles from the simulator have now been changed to be
defined as positive counter clockwise.