Update output_saver to use coordinate transform
Update the output_saver module to now use the new data coming from the
coordinate transform so that it now does not do the conversions itself.
Still a bit rough, and a lot of code is just commented out, but this
will be fixed later. It remains to update the visualization code in
order to be able to plot the new data, but the data saving seems to be
working as it is right now. Currently the part responsible for saving
the helipad ground truth has been commented out, as I don't know if that
will actually be needed at all since the drone is just moving relative
to the helipad anyways. Maybe later, it would be nice to have definitive
NED coordinates of both in addition to the relative drone measurement,
but I will find out that later.