Make visualization work with GT and DNN CV data
Synchronization also works, but is not very efficient as one function
now has to be made for each type of estimate. This also only allowes for
two datasets to be plottet in the same plot.

The orientation seems to work well. There is not much delay and the
estimates seem to follow the true values well. However, one small thing
is that when the angle flips from -180 degrees to 180 degrees because of
the SSA, it becomes zero at a brief instant inbetween the two. Not quite
sure why.