Update lab-test, add script to view img from topic
The lab test script now flies higher in order to see the helipad more.

The script to see the image is just to be able to see what the drone
sees when playing a rosbag with the images. This was tried to work for
the detections topic from darknet also, but this did not work and the
picture was always dark.