Change dnncv, coordinate trans. to track helipad
These now track the helipad in the drone body frame. This was done
because there is no way to do it in the helipad frame without knowing
the rotation of the helipad. It was thought to be cleaner to do it this
way anyways, as the position measurements should not be dependent on the
much less frequent heading measurements. These changes have been
imaplemented in the coordinate_transform.py and bb_pose_estmate.py
files, and the other files are changed in order to make them compatible
with the new changes.

A bug has been changed with the conversion of the angles coming from the
simulator. It was believed that these were defined positive clockwise
and from the North axis, but in fact they were defined positive counter
clockwise from the East axis. This has now been fixed, and a new
rotation matrix from NED to body has been made. This matrix is the same
as Fossen uses as from body to NED, but there is probably somewhere
where there is another definition or some angles or something that makes
them different.