Change EKF to use constant velocity model
Lots of files changed because they are all dependent on each other. This
new system now uses the constant velocity model and does not estimate
anything other than the position and velocity of the helipad relative to
the drone body frame. From one simple experiment it seems to be working
somewhat well, with the estimates for each axis following the actual
ones well. Tuning is probably not correct at all and it might be that
the estimates are just following the measurements completely.