Make EKF generic
Make the EKF module generic so that it can use different dynamic models
and measurement models. Now all the settings are changed in the
ekf_config.yaml config file and then the code handles setting up
everything based on that so that it should be easier to switch between
different filters, as well as adding measurements etc.