Update output_saver to save new EKF ouput
This is not that sophisticated and will have to be extended a bit to be
able to save output from different types of EKF filters with different
states etc.

A new message type was also created because it did not make sense to use
the Pose message from ROS because it was only outputing position, and
the covariance part of the message got the wrong amount of entries when
only supplied with the 9 covariances from the position.