Update visualization to plot EKF position  as well
Some progress towards making the visualization more general, but it is
still quite specific. Should be easier to get everything to work with
later.

Also added code to synchronize multiple data series and plot them in the
same plot. This should make it easier to assess the quality of the EKF
estimates for example.