Make ekf start after first measurement, add cov to plot
The EKF now starts after the first velocity measurement arrives so that
it is not predicting for an unneccesary amount of time before the other
systems start (the velocity measurement still starts very early).

Also added plotting of +-1 sigma of the EKF plot. This is very spiky as
it changes each time a new measurement arrives. This might be due to
improper tuning, or if not then it might have to be fixed by averaging
out the standard deviation over some small interval.