Fix bugs in EKF, tune EKF based on velocity meas.
The sigmas that are loaded into the EKF are now squared so that that
represent variance/covariance instead which they mistakenly did not do
earlier.

The CV model was also fixed to be correct, as it had the wrong signs in
the predict function f, and the prediction matrix F.

Tuning was performed using only the velocity measurements to try to get
this to work as well as possible. Seems to be working fairly well, with
the covariance roughly representing the actual error. Will now tune the
EKF based on the DNNCV measurements as well.