Try to fix weird velocity bug
Bug where velocity is positive in some parts of the simulation where it
should clearly be negative. This causes the estimates to diverge and
makes the estimate bad. If this can be changed then it seems like the
EKF could work very well.

Things that have been tried to fix the bug:
- Checked if there was something wrong with the conversion from NED to
body frame in the velocity coming from the drone interface by changing
the scipy rotations to the ned to body rotation defined in Fossen. This
did not change anything

Things to try:
- Check the raw ned velocity coming from the drone during the points of
interest and see if this is consistent with how it should be. If not
then it is a fault with the internal velocity estimation, and if it is
consistent then it is a fault with the conversion from NED to body
somewhere.