Add attitude control tests
Seems to work well enough in the simulator at least. Had a problem where
it sometimes ran flew much farther than it was supposed to, but this was
fixed by changing the queue_size of the published attitude setpoints to
1 to avoid a build-up of old messages in case the drone interface fail
to process all the messages at once. With only some limited testing this
seems to have fixed the problem, but have to watch out for it in the
future also.

Also, the rate at which the attitude setpoints are published when using
a rospy.Rate object is off in the first two messages, but then works
well. This is because it tries to keep a constant rate by computing how
long each of the elements of the loop takes. The fact that the first two
messages are published with too short of an interval shouldn't be a
problem.