Update output_saver to save velocity from mocap
The attitude is also being saved and this is just because the velocity
will be used to assess the attitude reference generator and therefore it
is nice to have both the ground truth velocity and attitude in the same
file with the same number of entries.

Has not been tested yet in the drone lab, so unsure if this works.