Save/visualize mocap vel/att, drone raw vel/att
Big commit with lots of changes. Main changes is that the velocity and
attitude from the odometry messages of the mocap system are now saved,
as well as the velocity and attitude coming straight from the drone.
This now lays the foundation for being able to create the attitude
reference generator because the data used for tuning the PID and
performing system identification is now available.

Also added some code to visualize the data that is now being saved. This
has to be improved later.