Update attitude setpoint to use correct pitch def
The pitch value that is supplied to the command listener is now
consisten with the roll-pitch-yaw definition of defining a postive pitch
angle as pointing the nose upwards. To get this to work with the
attitude setpoint from the PID controller, the gains for the x and y
axis had to be separated as the x-axis now requires a negative gain.