Add model based attitude ref with initial tuning
Implemented a model based attitude reference generator based on the
linear drag model used by Teppo Luukkonen in his paper "Modelling and
control of quadcopter".

The results are currently pretty good with some overshoot when
adding references to both vx and vy, but this should improve when
learing the parameters of the drag model better.

Have not implemented the reference generator based on the nonlinear drag
equation, as the linear one seems to work well and it may be easier to
learn the unknown parameters of.