Fix EKF bug with constant velocity model
The bug was that the model did not take into consideration any changes
in heading angle, and thus the estimates were completely off whenever
the drone changed heading. This was fixed by using the heading
measurement from the drone along with the previous measurement to create
a heading change which was then used to rotate the current measurement.
One flaw is that the heading measurements only arrives at 5Hz, meaning
that there are quite a few incorrect predictions inbetween the
measurements, but the overall predictions still followed the overall
shape of the ground truth data well.