Add PID control for tracking
Tested using just a P controller which worked well in simulations and in
the drone lab. In the drone lab I tried moving the helipad along the
floor and the drone was able to track it well. This was pretty cool to
see working.

The drone velocity measurements spike whenever I move the helipad
underneath it and this is likely due to optical flow being a part of the
internal EKF on the drone and that the are assuming that the ground does
not move. The measurements soon converge, but I don't know what would
happen if the helipad was continuously moving.