Add simple landing system
This system is meant as a demonstration that the trakcing if sufficient
to allow the drone to land on the helipad. It descends using the
internal position controller until it is at about 0.5m and then tracking
can be resumed until you want to land. There are three states: tracking,
descending and landing, and the transition between these are chosen by
the used via keyboare input.