Add experimental tcv and start of circle label
The experimental tcv file includes a lot of functions and a lot which
overlaps with other tcv files. This file is meant to be temporary and
has to be cleaned up and the functionality moved to respective files.
This is all the start of a larger effort to try to use machine learning
to learn the hyperparameters for the shi-tomasi corner detector and the
hough-circle circle detector.