Update TCV and make output saver compatible
The results using the TCV are so far not great. There are not a
estimates coming through, and not quite sure why this is the case.
Should be coming at least around 4-5 based on the time it takes to
calulate everything, but maybe there is something in the corner
identification that is rejecting too many points.